# apps/ecommerce-api/docs/contracts/openapi.ecommerce.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ProductAgentHandoffImportRequest": {
        "properties": {
          "catalog_source": {
            "default": "sourceCata",
            "title": "Catalog Source",
            "type": "string"
          },
          "file": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "File"
          },
          "file_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "File Path"
          },
          "limit": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Limit"
          },
          "persist_initial_capture": {
            "default": true,
            "title": "Persist Initial Capture",
            "type": "boolean"
          },
          "report_items_limit": {
            "default": 200,
            "title": "Report Items Limit",
            "type": "integer"
          }
        },
        "title": "ProductAgentHandoffImportRequest",
        "type": "object"
      },
...
      "SourceUrlImportRequest": {
        "properties": {
          "catalog_source": {
            "default": "sourceCata",
            "title": "Catalog Source",
            "type": "string"
          },
          "include_artifacts": {
            "default": true,
            "title": "Include Artifacts",
            "type": "boolean"
          },
          "include_legacy_runs": {
            "default": false,
            "title": "Include Legacy Runs",
            "type": "boolean"
          },
          "include_observations": {
            "default": true,
            "title": "Include Observations",
            "type": "boolean"
          },
          "legacy_runs_dir": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Legacy Runs Dir"
          },
          "limit": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Limit"
          },
          "report_items_limit": {
            "default": 200,
            "title": "Report Items Limit",
            "type": "integer"
          }
        },
        "title": "SourceUrlImportRequest",
        "type": "object"
      },
      "SourceUrlImportResponse": {
        "properties": {
          "applied": {
            "title": "Applied",
            "type": "boolean"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/SourceUrlImportCandidateReport"
            },
            "title": "Items",
            "type": "array"
          },
          "mode": {
            "title": "Mode",
            "type": "string"
          },
          "sources": {
            "additionalProperties": {
              "$ref": "#/components/schemas/SourceUrlImportSourceStats"
            },
            "title": "Sources",
            "type": "object"
          },
          "summary": {
            "$ref": "#/components/schemas/SourceUrlImportSummary"
          },
          "truncated": {
            "title": "Truncated",
            "type": "boolean"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "mode",
          "applied",
          "summary",
          "sources",
          "items",
          "truncated",
          "warnings"
        ],
        "title": "SourceUrlImportResponse",
        "type": "object"
      },
      "SourceUrlImportSourceStats": {
        "properties": {
          "candidates": {
            "default": 0,
            "title": "Candidates",
            "type": "integer"
          },
          "processed": {
            "default": 0,
            "title": "Processed",
            "type": "integer"
          }
        },
        "title": "SourceUrlImportSourceStats",
        "type": "object"
      },
...
      "SourceUrlSummaryResponse": {
        "properties": {
          "by_source_name": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "By Source Name",
            "type": "object"
          },
          "by_status": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "By Status",
            "type": "object"
          },
          "by_url_type": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "By Url Type",
            "type": "object"
          },
          "catalog_product_count": {
            "title": "Catalog Product Count",
            "type": "integer"
          },
          "catalog_source": {
            "title": "Catalog Source",
            "type": "string"
          },
          "coverage_percent": {
            "title": "Coverage Percent",
            "type": "number"
          },
          "products_with_active_source_urls": {
            "title": "Products With Active Source Urls",
            "type": "integer"
          },
          "products_without_active_source_urls": {
            "title": "Products Without Active Source Urls",
            "type": "integer"
          },
          "source_url_count": {
            "title": "Source Url Count",
            "type": "integer"
          },
          "updated_at": {
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "catalog_source",
          "catalog_product_count",
          "products_with_active_source_urls",
          "products_without_active_source_urls",
          "coverage_percent",
          "source_url_count",
          "by_status",
          "by_source_name",
          "by_url_type",
          "updated_at"
        ],
        "title": "SourceUrlSummaryResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Ecommerce API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/catalog/source-urls/import/apply": {
      "post": {
        "operationId": "apply_source_url_import_api_catalog_source_urls_import_apply_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SourceUrlImportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceUrlImportResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Apply Source Url Import",
        "tags": [
          "catalog-source-urls"
        ]
      }
    },
...
    "/api/catalog/source-urls/import/preview": {
      "post": {
        "operationId": "preview_source_url_import_api_catalog_source_urls_import_preview_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SourceUrlImportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceUrlImportResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Preview Source Url Import",
        "tags": [
          "catalog-source-urls"
        ]
      }
    },
    "/api/catalog/source-urls/import/product-agent/apply": {
      "post": {
        "operationId": "apply_product_agent_handoff_import_api_catalog_source_urls_import_product_agent_apply_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductAgentHandoffImportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceUrlImportResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Apply Product Agent Handoff Import",
        "tags": [
          "catalog-source-urls"
        ]
      }
    },
    "/api/catalog/source-urls/import/product-agent/preview": {
      "post": {
        "operationId": "preview_product_agent_handoff_import_api_catalog_source_urls_import_product_agent_preview_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductAgentHandoffImportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceUrlImportResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Preview Product Agent Handoff Import",
        "tags": [
          "catalog-source-urls"
        ]
      }
    },
    "/api/catalog/source-urls/summary": {
      "get": {
        "operationId": "get_source_url_summary_api_catalog_source_urls_summary_get",
        "parameters": [
          {
            "in": "query",
            "name": "catalog_source",
            "required": false,
            "schema": {
              "default": "sourceCata",
              "title": "Catalog Source",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceUrlSummaryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Source Url Summary",
        "tags": [
          "catalog-source-urls"
        ]
      }
    },
...
    "/api/vendor-sources/source-urls/summary": {
      "get": {
        "operationId": "get_vendor_source_url_summary_api_vendor_sources_source_urls_summary_get",
        "parameters": [
          {
            "in": "query",
            "name": "source_name",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source Name"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Vendor Source Url Summary Api Vendor Sources Source Urls Summary Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Vendor Source Url Summary",
        "tags": [
          "vendor-sources"
        ]
      }
    },
...
  }
}